This PR fakify ScriptObject inputs and attributes in export non-strict mode by default. 

The basic idea is to `only fakify the script object during tracing (i.e. aot_export)`. After we get the traced graph module, eagerly executing, serializing, or running more passes will use the real script objects. This is essentially treating the script object as constant tensor.

Concretely, we 
1. fakify all the script object inputs, and module attributes (gathered by constant_attrs). 
2. patch the module's attributes with fakified script object 
3. right after aot_export, remove the patching (to avoid changing the original module) then modify the exported graph module's attribute to real script object. 


[ghstack-poisoned]

+++ torch/csrc/distributed/c10d/TCPStore.hpp
@@ -30,10 +30,10 @@ class Counter {
     return count_;
   }
   double variance() const noexcept {
-    return m2_ / static_cast<double>(count_);
+    return m2_ / count_;
   }
   double sample_variance() const noexcept {
-    return m2_ / static_cast<double>(count_ - 1);
+    return m2_ / (count_ - 1);
   }
 
  private:
